backend/requirements.txt
fastapi==0.115.*
uvicorn==0.34.*
pvlib==0.14.*
pandas==2.2.*
numpy==2.1.*
shapely==2.0.*
scipy==1.14.*
openai==1.59.*
httpx==0.28.*
fal-client==0.5.*
python-multipart
websockets==12.*
python-dotenv
langgraph
langchain-openai
langchain-core
